Updated binary search  java as requested
Removed sorting of arrays after taking input.v because its a binary search it works only on sorted arrays. Its of no use on other arrays as it will then affect its time complexity as well as results. So asking user initially to enter the array only in increasing order.